// my-app/src/app/services/contacto/contacto-interface.ts
export interface ContactoInterface {
    id?: number;
    name?: string;
    surname?: string;
    town_id?: number;
    phone_p?: string;
    phone_s?: string;
    sex?: string;
    mail?: string;
    institution?: string;
    office_id?: number;
    user_id?: number;
    created_at?: string;
    updated_at?: string;
}